Move menus and buttons into subdirectory
