+++ src/components/editor/document-bar/buttons/share-link-button.tsx
@@ -1,9 +1,9 @@
 import React, { Fragment, useState } from 'react'
 import { Button, Modal } from 'react-bootstrap'
 import { Trans } from 'react-i18next'
-import { CopyableField } from '../../common/copyable-field/copyable-field'
-import { ForkAwesomeIcon } from '../../common/fork-awesome/fork-awesome-icon'
-import { CommonModal } from '../../common/modals/common-modal'
+import { CopyableField } from '../../../common/copyable-field/copyable-field'
+import { ForkAwesomeIcon } from '../../../common/fork-awesome/fork-awesome-icon'
+import { CommonModal } from '../../../common/modals/common-modal'
 
 export const ShareLinkButton: React.FC = () => {
   const [showReadOnly, setShowReadOnly] = useState(false)
