+++ src/components/editor/document-bar/menus/editor-menu.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Dropdown } from 'react-bootstrap'
 import { Trans, useTranslation } from 'react-i18next'
-import { DropdownItemWithDeletionModal } from '../../history-page/entry-menu/dropdown-item-with-deletion-modal'
+import { DropdownItemWithDeletionModal } from '../../../history-page/entry-menu/dropdown-item-with-deletion-modal'
 
 export interface EditorMenuProps {
   noteTitle: string
